feat(performance): improves the load time performance of the modal drastically.
Description:

Currently the modal makes calls to the REST end point every time it is being opened.
As a result, this takes roughly 5 seconds to process the response.

This PR handles that by doing the call on the background just once and cache the response for a specified input set as it won't ever change for that particular input set. This enables the modal to load instantaneously like a static modal.
This improves the users' experience as well.

Whenever the input set changes, a call is made to fetch the changed response which is then cached.

Tracks:

https://github.com/openshiftio/openshift.io/issues/3863
https://github.com/openshiftio/openshift.io/issues/3885